requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
sqlalchemy
ibm-watson>=8.0.0
